+++ src/main/java/org/kumoricon/site/attendee/AttendeePrintView.java
@@ -1,6 +1,8 @@
 package org.kumoricon.site.attendee;
 
+import com.vaadin.ui.Window;
 import org.kumoricon.model.attendee.Attendee;
+import org.kumoricon.model.user.User;
 
 import java.util.List;
 
@@ -10,4 +12,12 @@ public interface AttendeePrintView {
     void notifyError(String message);
     String getCurrentUsername();
     String getCurrentClientIPAddress();
+
+    Boolean currentUserHasRight(String pre_reg_check_in);
+
+    void showWindow(Window window);
+
+    User getCurrentUser();
+
+    void refresh();
 }