+++ src/main/java/org/kumoricon/site/attendee/CheckInConfirmationHandler.java
@@ -1,0 +1,5 @@
+package org.kumoricon.site.attendee;
+
+public interface CheckInConfirmationHandler {
+    void attendeeInformationVerified();
+}